Computes all the chosen transformations to the input data frame in the chosen parameters.

Computes all the chosen transformations to the input data frame in the chosen parameters.

alltransf(
  data,
  subset,
  select = colnames(data),
  transf = c("rsqrt", "log", "sqrt")
)

Arguments

data
A data frame

subset
a specification of the rows to be used: defaults to all rows. This can be any valid indexing vector (see [.data.frame) for the rows of data or if that is not supplied, a data frame made up of the variables used in formula.

select
expression, indicating columns to select from a data frame. Defaults for all the variables in data. See subset.

transf
The chosen transformations to be applied to the data

Value

A list with all data transformed according to the arguments passed.help

Examples

library(appraiseR)
data(centro_2015)
dados <- centro_2015
vars <- colnames(dados)
alltransf(dados, select = vars)

#> $identity
#>      valor area_total quartos suites garagens dist_b_mar
#> 1  1060000     350.00       3      1        2        720
#> 2   510000     136.56       3      1        1        665
#> 3   780000     164.77       3      1        2        415
#> 4   550000     174.58       3      1        1        320
#> 5   850000     123.01       3      1        3        895
#> 6   300000      89.83       2      0        1        645
#> 7   750000     174.00       2      1        2        860
#> 8   650000     123.00       3      1        1        745
#> 9   620000     121.00       3      1        1        745
#> 10  740000     109.00       3      1        1        300
#> 11  770000     170.00       3      1        2        590
#> 12  680000     141.00       3      1        1        290
#> 13  850000     174.00       3      1        1        465
#> 14  420000     105.00       3      1        0         60
#> 15  547000     128.00       3      1        1        745
#> 16 1600000     163.00       4      2        2         90
#> 17 1320000     230.00       3      1        2        215
#> 18  615000     108.00       3      1        1        745
#> 19  705000     174.00       2      1        2        900
#> 20  418000      85.00       1      0        1        620
#> 21  270000      71.00       2      0        0       1380
#> 22  418000     100.00       1      1        1        620
#> 23  650000      90.00       2      1        1        215
#> 24  700000     161.00       2      1        2        500
#> 25  680000     174.00       2      1        2        860
#> 26  420000      76.00       2      1        1        700
#> 27  195000      48.00       1      0        0        730
#> 28  290000      66.00       1      0        1        745
#> 29  272000      50.00       1      0        1       1430
#> 30  430000      61.00       2      0        1        170
#> 31  895000     109.00       3      1        1        530
#> 32  450000      89.00       2      0        1        745
#> 33 1950000     393.00       3      1        3        550
#> 34 2150000     578.00       3      2        3        260
#> 35  940000     182.00       3      1        2        200
#> 36 1400000     262.00       4      1        1         60
#> 37 1090000     205.00       3      0        3        465
#> 38 1272000     196.00       3      3        2        610
#> 39 2800000     463.00       3      3        3        590
#> 40 1796000     273.00       3      3        4        140
#> 41 1400000     330.00       4      2        2        655
#> 42 3000000     533.00       4      3        4        427
#> 43 1200000     221.00       3      3        2        607
#> 44  800000     220.00       3      1        1       1000
#> 45  950000     127.00       2      1        1         60
#> 46 2061000     362.00       3      3        4        310
#> 47 1326000     315.00       3      3        3        600
#> 48  850000     151.00       3      1        2        660
#> 49 1650000     246.00       3      3        3        307
#> 50  650000     159.72       3      1        1        120
#> 
#> $rsqrt
#>           valor area_total   quartos    suites  garagens dist_b_mar
#> 1  0.0009712859 0.05345225 0.5773503 1.0000000 0.7071068 0.03726780
#> 2  0.0014002801 0.08557329 0.5773503 1.0000000 1.0000000 0.03877834
#> 3  0.0011322770 0.07790421 0.5773503 1.0000000 0.7071068 0.04908807
#> 4  0.0013483997 0.07568377 0.5773503 1.0000000 1.0000000 0.05590170
#> 5  0.0010846523 0.09016330 0.5773503 1.0000000 0.5773503 0.03342631
#> 6  0.0018257419 0.10550895 0.7071068       Inf 1.0000000 0.03937496
#> 7  0.0011547005 0.07580980 0.7071068 1.0000000 0.7071068 0.03409972
#> 8  0.0012403473 0.09016696 0.5773503 1.0000000 1.0000000 0.03663717
#> 9  0.0012700013 0.09090909 0.5773503 1.0000000 1.0000000 0.03663717
#> 10 0.0011624764 0.09578263 0.5773503 1.0000000 1.0000000 0.05773503
#> 11 0.0011396058 0.07669650 0.5773503 1.0000000 0.7071068 0.04116935
#> 12 0.0012126781 0.08421519 0.5773503 1.0000000 1.0000000 0.05872202
#> 13 0.0010846523 0.07580980 0.5773503 1.0000000 1.0000000 0.04637389
#> 14 0.0015430335 0.09759001 0.5773503 1.0000000       Inf 0.12909944
#> 15 0.0013520923 0.08838835 0.5773503 1.0000000 1.0000000 0.03663717
#> 16 0.0007905694 0.07832604 0.5000000 0.7071068 0.7071068 0.10540926
#> 17 0.0008703883 0.06593805 0.5773503 1.0000000 0.7071068 0.06819943
#> 18 0.0012751534 0.09622504 0.5773503 1.0000000 1.0000000 0.03663717
#> 19 0.0011909827 0.07580980 0.7071068 1.0000000 0.7071068 0.03333333
#> 20 0.0015467206 0.10846523 1.0000000       Inf 1.0000000 0.04016097
#> 21 0.0019245009 0.11867817 0.7071068       Inf       Inf 0.02691910
#> 22 0.0015467206 0.10000000 1.0000000 1.0000000 1.0000000 0.04016097
#> 23 0.0012403473 0.10540926 0.7071068 1.0000000 1.0000000 0.06819943
#> 24 0.0011952286 0.07881104 0.7071068 1.0000000 0.7071068 0.04472136
#> 25 0.0012126781 0.07580980 0.7071068 1.0000000 0.7071068 0.03409972
#> 26 0.0015430335 0.11470787 0.7071068 1.0000000 1.0000000 0.03779645
#> 27 0.0022645541 0.14433757 1.0000000       Inf       Inf 0.03701166
#> 28 0.0018569534 0.12309149 1.0000000       Inf 1.0000000 0.03663717
#> 29 0.0019174125 0.14142136 1.0000000       Inf 1.0000000 0.02644429
#> 30 0.0015249857 0.12803688 0.7071068       Inf 1.0000000 0.07669650
#> 31 0.0010570328 0.09578263 0.5773503 1.0000000 1.0000000 0.04343722
#> 32 0.0014907120 0.10599979 0.7071068       Inf 1.0000000 0.03663717
#> 33 0.0007161149 0.05044333 0.5773503 1.0000000 0.5773503 0.04264014
#> 34 0.0006819943 0.04159452 0.5773503 0.7071068 0.5773503 0.06201737
#> 35 0.0010314212 0.07412493 0.5773503 1.0000000 0.7071068 0.07071068
#> 36 0.0008451543 0.06178021 0.5000000 1.0000000 1.0000000 0.12909944
#> 37 0.0009578263 0.06984303 0.5773503       Inf 0.5773503 0.04637389
#> 38 0.0008866586 0.07142857 0.5773503 0.5773503 0.7071068 0.04048882
#> 39 0.0005976143 0.04647394 0.5773503 0.5773503 0.5773503 0.04116935
#> 40 0.0007461855 0.06052275 0.5773503 0.5773503 0.5000000 0.08451543
#> 41 0.0008451543 0.05504819 0.5000000 0.7071068 0.7071068 0.03907323
#> 42 0.0005773503 0.04331481 0.5000000 0.5773503 0.5000000 0.04839339
#> 43 0.0009128709 0.06726728 0.5773503 0.5773503 0.7071068 0.04058875
#> 44 0.0011180340 0.06741999 0.5773503 1.0000000 1.0000000 0.03162278
#> 45 0.0010259784 0.08873565 0.7071068 1.0000000 1.0000000 0.12909944
#> 46 0.0006965640 0.05255883 0.5773503 0.5773503 0.5000000 0.05679618
#> 47 0.0008684168 0.05634362 0.5773503 0.5773503 0.5773503 0.04082483
#> 48 0.0010846523 0.08137885 0.5773503 1.0000000 0.7071068 0.03892495
#> 49 0.0007784989 0.06375767 0.5773503 0.5773503 0.5773503 0.05707301
#> 50 0.0012403473 0.07912621 0.5773503 1.0000000 1.0000000 0.09128709
#> 
#> $log
#>       valor area_total   quartos    suites  garagens dist_b_mar
#> 1  13.87378   5.857933 1.0986123 0.0000000 0.6931472   6.579251
#> 2  13.14217   4.916764 1.0986123 0.0000000 0.0000000   6.499787
#> 3  13.56705   5.104551 1.0986123 0.0000000 0.6931472   6.028279
#> 4  13.21767   5.162383 1.0986123 0.0000000 0.0000000   5.768321
#> 5  13.65299   4.812266 1.0986123 0.0000000 1.0986123   6.796824
#> 6  12.61154   4.497919 0.6931472      -Inf 0.0000000   6.469250
#> 7  13.52783   5.159055 0.6931472 0.0000000 0.6931472   6.756932
#> 8  13.38473   4.812184 1.0986123 0.0000000 0.0000000   6.613384
#> 9  13.33747   4.795791 1.0986123 0.0000000 0.0000000   6.613384
#> 10 13.51441   4.691348 1.0986123 0.0000000 0.0000000   5.703782
#> 11 13.55415   5.135798 1.0986123 0.0000000 0.6931472   6.380123
#> 12 13.42985   4.948760 1.0986123 0.0000000 0.0000000   5.669881
#> 13 13.65299   5.159055 1.0986123 0.0000000 0.0000000   6.142037
#> 14 12.94801   4.653960 1.0986123 0.0000000      -Inf   4.094345
#> 15 13.21220   4.852030 1.0986123 0.0000000 0.0000000   6.613384
#> 16 14.28551   5.093750 1.3862944 0.6931472 0.6931472   4.499810
#> 17 14.09314   5.438079 1.0986123 0.0000000 0.6931472   5.370638
#> 18 13.32938   4.682131 1.0986123 0.0000000 0.0000000   6.613384
#> 19 13.46595   5.159055 0.6931472 0.0000000 0.6931472   6.802395
#> 20 12.94324   4.442651 0.0000000      -Inf 0.0000000   6.429719
#> 21 12.50618   4.262680 0.6931472      -Inf      -Inf   7.229839
#> 22 12.94324   4.605170 0.0000000 0.0000000 0.0000000   6.429719
#> 23 13.38473   4.499810 0.6931472 0.0000000 0.0000000   5.370638
#> 24 13.45884   5.081404 0.6931472 0.0000000 0.6931472   6.214608
#> 25 13.42985   5.159055 0.6931472 0.0000000 0.6931472   6.756932
#> 26 12.94801   4.330733 0.6931472 0.0000000 0.0000000   6.551080
#> 27 12.18075   3.871201 0.0000000      -Inf      -Inf   6.593045
#> 28 12.57764   4.189655 0.0000000      -Inf 0.0000000   6.613384
#> 29 12.51356   3.912023 0.0000000      -Inf 0.0000000   7.265430
#> 30 12.97154   4.110874 0.6931472      -Inf 0.0000000   5.135798
#> 31 13.70458   4.691348 1.0986123 0.0000000 0.0000000   6.272877
#> 32 13.01700   4.488636 0.6931472      -Inf 0.0000000   6.613384
#> 33 14.48334   5.973810 1.0986123 0.0000000 1.0986123   6.309918
#> 34 14.58098   6.359574 1.0986123 0.6931472 1.0986123   5.560682
#> 35 13.75364   5.204007 1.0986123 0.0000000 0.6931472   5.298317
#> 36 14.15198   5.568345 1.3862944 0.0000000 0.0000000   4.094345
#> 37 13.90169   5.323010 1.0986123      -Inf 1.0986123   6.142037
#> 38 14.05610   5.278115 1.0986123 1.0986123 0.6931472   6.413459
#> 39 14.84513   6.137727 1.0986123 1.0986123 1.0986123   6.380123
#> 40 14.40107   5.609472 1.0986123 1.0986123 1.3862944   4.941642
#> 41 14.15198   5.799093 1.3862944 0.6931472 0.6931472   6.484635
#> 42 14.91412   6.278521 1.3862944 1.0986123 1.3862944   6.056784
#> 43 13.99783   5.398163 1.0986123 1.0986123 0.6931472   6.408529
#> 44 13.59237   5.393628 1.0986123 0.0000000 0.0000000   6.907755
#> 45 13.76422   4.844187 0.6931472 0.0000000 0.0000000   4.094345
#> 46 14.53870   5.891644 1.0986123 1.0986123 1.3862944   5.736572
#> 47 14.09768   5.752573 1.0986123 1.0986123 1.0986123   6.396930
#> 48 13.65299   5.017280 1.0986123 0.0000000 0.6931472   6.492240
#> 49 14.31629   5.505332 1.0986123 1.0986123 1.0986123   5.726848
#> 50 13.38473   5.073422 1.0986123 0.0000000 0.0000000   4.787492
#> 
#> $sqrt
#>        valor area_total  quartos   suites garagens dist_b_mar
#> 1  1029.5630  18.708287 1.732051 1.000000 1.414214  26.832816
#> 2   714.1428  11.685889 1.732051 1.000000 1.000000  25.787594
#> 3   883.1761  12.836277 1.732051 1.000000 1.414214  20.371549
#> 4   741.6198  13.212873 1.732051 1.000000 1.000000  17.888544
#> 5   921.9544  11.090987 1.732051 1.000000 1.732051  29.916551
#> 6   547.7226   9.477869 1.414214 0.000000 1.000000  25.396850
#> 7   866.0254  13.190906 1.414214 1.000000 1.414214  29.325757
#> 8   806.2258  11.090537 1.732051 1.000000 1.000000  27.294688
#> 9   787.4008  11.000000 1.732051 1.000000 1.000000  27.294688
#> 10  860.2325  10.440307 1.732051 1.000000 1.000000  17.320508
#> 11  877.4964  13.038405 1.732051 1.000000 1.414214  24.289916
#> 12  824.6211  11.874342 1.732051 1.000000 1.000000  17.029386
#> 13  921.9544  13.190906 1.732051 1.000000 1.000000  21.563859
#> 14  648.0741  10.246951 1.732051 1.000000 0.000000   7.745967
#> 15  739.5945  11.313708 1.732051 1.000000 1.000000  27.294688
#> 16 1264.9111  12.767145 2.000000 1.414214 1.414214   9.486833
#> 17 1148.9125  15.165751 1.732051 1.000000 1.414214  14.662878
#> 18  784.2194  10.392305 1.732051 1.000000 1.000000  27.294688
#> 19  839.6428  13.190906 1.414214 1.000000 1.414214  30.000000
#> 20  646.5292   9.219544 1.000000 0.000000 1.000000  24.899799
#> 21  519.6152   8.426150 1.414214 0.000000 0.000000  37.148351
#> 22  646.5292  10.000000 1.000000 1.000000 1.000000  24.899799
#> 23  806.2258   9.486833 1.414214 1.000000 1.000000  14.662878
#> 24  836.6600  12.688578 1.414214 1.000000 1.414214  22.360680
#> 25  824.6211  13.190906 1.414214 1.000000 1.414214  29.325757
#> 26  648.0741   8.717798 1.414214 1.000000 1.000000  26.457513
#> 27  441.5880   6.928203 1.000000 0.000000 0.000000  27.018512
#> 28  538.5165   8.124038 1.000000 0.000000 1.000000  27.294688
#> 29  521.5362   7.071068 1.000000 0.000000 1.000000  37.815341
#> 30  655.7439   7.810250 1.414214 0.000000 1.000000  13.038405
#> 31  946.0444  10.440307 1.732051 1.000000 1.000000  23.021729
#> 32  670.8204   9.433981 1.414214 0.000000 1.000000  27.294688
#> 33 1396.4240  19.824228 1.732051 1.000000 1.732051  23.452079
#> 34 1466.2878  24.041631 1.732051 1.414214 1.732051  16.124515
#> 35  969.5360  13.490738 1.732051 1.000000 1.414214  14.142136
#> 36 1183.2160  16.186414 2.000000 1.000000 1.000000   7.745967
#> 37 1044.0307  14.317821 1.732051 0.000000 1.732051  21.563859
#> 38 1127.8298  14.000000 1.732051 1.732051 1.414214  24.698178
#> 39 1673.3201  21.517435 1.732051 1.732051 1.732051  24.289916
#> 40 1340.1492  16.522712 1.732051 1.732051 2.000000  11.832160
#> 41 1183.2160  18.165902 2.000000 1.414214 1.414214  25.592968
#> 42 1732.0508  23.086793 2.000000 1.732051 2.000000  20.663978
#> 43 1095.4451  14.866069 1.732051 1.732051 1.414214  24.637370
#> 44  894.4272  14.832397 1.732051 1.000000 1.000000  31.622777
#> 45  974.6794  11.269428 1.414214 1.000000 1.000000   7.745967
#> 46 1435.6183  19.026298 1.732051 1.732051 2.000000  17.606817
#> 47 1151.5207  17.748239 1.732051 1.732051 1.732051  24.494897
#> 48  921.9544  12.288206 1.732051 1.000000 1.414214  25.690465
#> 49 1284.5233  15.684387 1.732051 1.732051 1.732051  17.521415
#> 50  806.2258  12.638038 1.732051 1.000000 1.000000  10.954451
#> 

alltransf(dados, 1:10, c("valor", "area_total"))

#> $identity
#>      valor area_total
#> 1  1060000     350.00
#> 2   510000     136.56
#> 3   780000     164.77
#> 4   550000     174.58
#> 5   850000     123.01
#> 6   300000      89.83
#> 7   750000     174.00
#> 8   650000     123.00
#> 9   620000     121.00
#> 10  740000     109.00
#> 
#> $rsqrt
#>           valor area_total
#> 1  0.0009712859 0.05345225
#> 2  0.0014002801 0.08557329
#> 3  0.0011322770 0.07790421
#> 4  0.0013483997 0.07568377
#> 5  0.0010846523 0.09016330
#> 6  0.0018257419 0.10550895
#> 7  0.0011547005 0.07580980
#> 8  0.0012403473 0.09016696
#> 9  0.0012700013 0.09090909
#> 10 0.0011624764 0.09578263
#> 
#> $log
#>       valor area_total
#> 1  13.87378   5.857933
#> 2  13.14217   4.916764
#> 3  13.56705   5.104551
#> 4  13.21767   5.162383
#> 5  13.65299   4.812266
#> 6  12.61154   4.497919
#> 7  13.52783   5.159055
#> 8  13.38473   4.812184
#> 9  13.33747   4.795791
#> 10 13.51441   4.691348
#> 
#> $sqrt
#>        valor area_total
#> 1  1029.5630  18.708287
#> 2   714.1428  11.685889
#> 3   883.1761  12.836277
#> 4   741.6198  13.212873
#> 5   921.9544  11.090987
#> 6   547.7226   9.477869
#> 7   866.0254  13.190906
#> 8   806.2258  11.090537
#> 9   787.4008  11.000000
#> 10  860.2325  10.440307
#>