Based on a dataframe containing the estimated values, returns amplitude according to NBR14.653-2

Amplitude according to NBR14.653-2.

Precision degree based on NBR14.653-2.

Interval of arbitration according to NBR14.653-2.

Interval evaluation based on NBR14.653-2.

amplitude(Y)

g_precisao(amplitude)

campo_arbitrio(Y)

aval_intervalar(Y)

Arguments

Y
dataframe with Y_hat values with confidence/prediction bestfit

amplitude
a vector containing the amplitudes of the estimations bestfit

Value

a vector containing the amplitudes of the Y_hat in %

a vector containing the precision degrees of the estimations.

a vector containing the interval of arbitration according to NBR14.653-2.

a vector containing interval evaluation according to NBR14.653-2.

Examples

fit <- lm(log(valor) ~ ., data = centro_2015@data)
require(dplyr)
aval <- centro_2015@data %>% filter(is.na(valor))
Y_hat <- predict(object = fit, interval = "confidence", newdata = aval)
Y_hat <- inverse(Y_hat, "log")
amp <- amplitude(Y_hat)
amp

#>     1     2     3 
#> 34.21 18.48 21.17 

g_precisao(amp)

#>     1     2     3 
#>  "II" "III" "III" 

campo_arbitrio(Y_hat)

#>     C.A.I.    C.A.S.
#> 1 637358.6  862308.6
#> 2 833530.0 1127717.1
#> 3 889919.5 1204008.7

aval_intervalar(Y_hat)

#>       L.I.      L.S.
#> 1 637358.6  862308.6
#> 2 894177.6 1075426.7
#> 3 942009.7 1163612.1