Summary method for class bestfit

summary method for class bestfit produce result summary containing the formula for the best (chosen) fit and the summary.lm for that fit.

# S3 method for bestfit
summary(object, fit = 1, subset, ...)

Arguments

object
an object of class bestfit.

fit
the number of the chosen fit from the combination matrix (defaults for the best fit found with bestfit).

subset
a specification of the rows to be used: defaults to all rows. This can be any valid indexing vector (see [.data.frame) for the rows of data or if that is not supplied, a data frame made up of the variables used in formula.

...
further arguments passed to lm.

Value

Returns the call for the bestfit function, the best (chosen) fit number, the lm formula and the lm fit summary for the best (chosen) fit transformations found by bestfit.

Examples

library(appraiseR)
dados <- centro_2015@data
best_fit <- bestfit(valor ~ ., data = dados)
summary(best_fit)

#> Error in is.data.frame(data): objeto 'dados' não encontrado

summary(best_fit, fit = 2)

#> Error in is.data.frame(data): objeto 'dados' não encontrado