Computes all the chosen transformations to the input data frame in the chosen parameters.

Computes all the chosen transformations to the input data frame in the chosen parameters.

alltransf(data, subset, select = colnames(data), transf = c("rsqrt", "log",
  "sqrt"))

Arguments

data
A data frame

subset
a specification of the rows to be used: defaults to all rows. This can be any valid indexing vector (see [.data.frame) for the rows of data or if that is not supplied, a data frame made up of the variables used in formula.

select
expression, indicating columns to select from a data frame. Defaults for all the variables in data. See subset.

transf
The chosen transformations to be applied to the data

Value

A list with all data transformed according to the arguments passed.help

Examples

vars <- colnames(centro_2015@data)
alltransf(centro_2015@data, select = vars)

#> $identity
#>      valor area_total quartos suites garagens dist_b_mar
#> 1  1060000     350.00       3      1        2        720
#> 2   510000     136.56       3      1        1        665
#> 3   780000     164.77       3      1        2        415
#> 4   550000     174.58       3      1        1        320
#> 5   850000     123.01       3      1        3        895
#> 6   300000      89.83       2      0        1        645
#> 7   750000     174.00       2      1        2        860
#> 8   650000     123.00       3      1        1        745
#> 9   620000     121.00       3      1        1        745
#> 10  740000     109.00       3      1        1        300
#> 11  770000     170.00       3      1        2        590
#> 12  680000     141.00       3      1        1        290
#> 13  850000     174.00       3      1        1        465
#> 14  420000     105.00       3      1        0         60
#> 15  547000     128.00       3      1        1        745
#> 16 1600000     163.00       4      2        2         90
#> 17 1320000     230.00       3      1        2        215
#> 18  615000     108.00       3      1        1        745
#> 19  705000     174.00       2      1        2        900
#> 20  418000      85.00       1      0        1        620
#> 21  270000      71.00       2      0        0       1380
#> 22  418000     100.00       1      1        1        620
#> 23  650000      90.00       2      1        1        215
#> 24  700000     161.00       2      1        2        500
#> 25  680000     174.00       2      1        2        860
#> 26  420000      76.00       2      1        1        700
#> 27  195000      48.00       1      0        0        730
#> 28  290000      66.00       1      0        1        745
#> 29  272000      50.00       1      0        1       1430
#> 30  430000      61.00       2      0        1        170
#> 31  895000     109.00       3      1        1        530
#> 32  450000      89.00       2      0        1        745
#> 33 1950000     393.00       3      1        3        550
#> 34 2150000     578.00       3      2        3        260
#> 35  940000     182.00       3      1        2        200
#> 36 1400000     262.00       4      1        1         60
#> 37 1090000     205.00       3      0        3        465
#> 38 1272000     196.00       3      3        2        610
#> 39 2800000     463.00       3      3        3        590
#> 40 1796000     273.00       3      3        4        140
#> 41 1400000     330.00       4      2        2        655
#> 42 3000000     533.00       4      3        4        427
#> 43 1200000     221.00       3      3        2        607
#> 44  800000     220.00       3      1        1       1000
#> 45  950000     127.00       2      1        1         60
#> 46 2061000     362.00       3      3        4        310
#> 47 1326000     315.00       3      3        3        600
#> 48  850000     151.00       3      1        2        660
#> 49 1650000     246.00       3      3        3        307
#> 50  650000     159.72       3      1        1        120
#> 
#> $rsqrt
#>           valor area_total   quartos    suites  garagens dist_b_mar
#> 1  0.0009712859 0.05345225 0.5773503 1.0000000 0.7071068 0.03726780
#> 2  0.0014002801 0.08557329 0.5773503 1.0000000 1.0000000 0.03877834
#> 3  0.0011322770 0.07790421 0.5773503 1.0000000 0.7071068 0.04908807
#> 4  0.0013483997 0.07568377 0.5773503 1.0000000 1.0000000 0.05590170
#> 5  0.0010846523 0.09016330 0.5773503 1.0000000 0.5773503 0.03342631
#> 6  0.0018257419 0.10550895 0.7071068       Inf 1.0000000 0.03937496
#> 7  0.0011547005 0.07580980 0.7071068 1.0000000 0.7071068 0.03409972
#> 8  0.0012403473 0.09016696 0.5773503 1.0000000 1.0000000 0.03663717
#> 9  0.0012700013 0.09090909 0.5773503 1.0000000 1.0000000 0.03663717
#> 10 0.0011624764 0.09578263 0.5773503 1.0000000 1.0000000 0.05773503
#> 11 0.0011396058 0.07669650 0.5773503 1.0000000 0.7071068 0.04116935
#> 12 0.0012126781 0.08421519 0.5773503 1.0000000 1.0000000 0.05872202
#> 13 0.0010846523 0.07580980 0.5773503 1.0000000 1.0000000 0.04637389
#> 14 0.0015430335 0.09759001 0.5773503 1.0000000       Inf 0.12909944
#> 15 0.0013520923 0.08838835 0.5773503 1.0000000 1.0000000 0.03663717
#> 16 0.0007905694 0.07832604 0.5000000 0.7071068 0.7071068 0.10540926
#> 17 0.0008703883 0.06593805 0.5773503 1.0000000 0.7071068 0.06819943
#> 18 0.0012751534 0.09622504 0.5773503 1.0000000 1.0000000 0.03663717
#> 19 0.0011909827 0.07580980 0.7071068 1.0000000 0.7071068 0.03333333
#> 20 0.0015467206 0.10846523 1.0000000       Inf 1.0000000 0.04016097
#> 21 0.0019245009 0.11867817 0.7071068       Inf       Inf 0.02691910
#> 22 0.0015467206 0.10000000 1.0000000 1.0000000 1.0000000 0.04016097
#> 23 0.0012403473 0.10540926 0.7071068 1.0000000 1.0000000 0.06819943
#> 24 0.0011952286 0.07881104 0.7071068 1.0000000 0.7071068 0.04472136
#> 25 0.0012126781 0.07580980 0.7071068 1.0000000 0.7071068 0.03409972
#> 26 0.0015430335 0.11470787 0.7071068 1.0000000 1.0000000 0.03779645
#> 27 0.0022645541 0.14433757 1.0000000       Inf       Inf 0.03701166
#> 28 0.0018569534 0.12309149 1.0000000       Inf 1.0000000 0.03663717
#> 29 0.0019174125 0.14142136 1.0000000       Inf 1.0000000 0.02644429
#> 30 0.0015249857 0.12803688 0.7071068       Inf 1.0000000 0.07669650
#> 31 0.0010570328 0.09578263 0.5773503 1.0000000 1.0000000 0.04343722
#> 32 0.0014907120 0.10599979 0.7071068       Inf 1.0000000 0.03663717
#> 33 0.0007161149 0.05044333 0.5773503 1.0000000 0.5773503 0.04264014
#> 34 0.0006819943 0.04159452 0.5773503 0.7071068 0.5773503 0.06201737
#> 35 0.0010314212 0.07412493 0.5773503 1.0000000 0.7071068 0.07071068
#> 36 0.0008451543 0.06178021 0.5000000 1.0000000 1.0000000 0.12909944
#> 37 0.0009578263 0.06984303 0.5773503       Inf 0.5773503 0.04637389
#> 38 0.0008866586 0.07142857 0.5773503 0.5773503 0.7071068 0.04048882
#> 39 0.0005976143 0.04647394 0.5773503 0.5773503 0.5773503 0.04116935
#> 40 0.0007461855 0.06052275 0.5773503 0.5773503 0.5000000 0.08451543
#> 41 0.0008451543 0.05504819 0.5000000 0.7071068 0.7071068 0.03907323
#> 42 0.0005773503 0.04331481 0.5000000 0.5773503 0.5000000 0.04839339
#> 43 0.0009128709 0.06726728 0.5773503 0.5773503 0.7071068 0.04058875
#> 44 0.0011180340 0.06741999 0.5773503 1.0000000 1.0000000 0.03162278
#> 45 0.0010259784 0.08873565 0.7071068 1.0000000 1.0000000 0.12909944
#> 46 0.0006965640 0.05255883 0.5773503 0.5773503 0.5000000 0.05679618
#> 47 0.0008684168 0.05634362 0.5773503 0.5773503 0.5773503 0.04082483
#> 48 0.0010846523 0.08137885 0.5773503 1.0000000 0.7071068 0.03892495
#> 49 0.0007784989 0.06375767 0.5773503 0.5773503 0.5773503 0.05707301
#> 50 0.0012403473 0.07912621 0.5773503 1.0000000 1.0000000 0.09128709
#> 
#> $log
#>       valor area_total   quartos    suites  garagens dist_b_mar
#> 1  13.87378   5.857933 1.0986123 0.0000000 0.6931472   6.579251
#> 2  13.14217   4.916764 1.0986123 0.0000000 0.0000000   6.499787
#> 3  13.56705   5.104551 1.0986123 0.0000000 0.6931472   6.028279
#> 4  13.21767   5.162383 1.0986123 0.0000000 0.0000000   5.768321
#> 5  13.65299   4.812266 1.0986123 0.0000000 1.0986123   6.796824
#> 6  12.61154   4.497919 0.6931472      -Inf 0.0000000   6.469250
#> 7  13.52783   5.159055 0.6931472 0.0000000 0.6931472   6.756932
#> 8  13.38473   4.812184 1.0986123 0.0000000 0.0000000   6.613384
#> 9  13.33747   4.795791 1.0986123 0.0000000 0.0000000   6.613384
#> 10 13.51441   4.691348 1.0986123 0.0000000 0.0000000   5.703782
#> 11 13.55415   5.135798 1.0986123 0.0000000 0.6931472   6.380123
#> 12 13.42985   4.948760 1.0986123 0.0000000 0.0000000   5.669881
#> 13 13.65299   5.159055 1.0986123 0.0000000 0.0000000   6.142037
#> 14 12.94801   4.653960 1.0986123 0.0000000      -Inf   4.094345
#> 15 13.21220   4.852030 1.0986123 0.0000000 0.0000000   6.613384
#> 16 14.28551   5.093750 1.3862944 0.6931472 0.6931472   4.499810
#> 17 14.09314   5.438079 1.0986123 0.0000000 0.6931472   5.370638
#> 18 13.32938   4.682131 1.0986123 0.0000000 0.0000000   6.613384
#> 19 13.46595   5.159055 0.6931472 0.0000000 0.6931472   6.802395
#> 20 12.94324   4.442651 0.0000000      -Inf 0.0000000   6.429719
#> 21 12.50618   4.262680 0.6931472      -Inf      -Inf   7.229839
#> 22 12.94324   4.605170 0.0000000 0.0000000 0.0000000   6.429719
#> 23 13.38473   4.499810 0.6931472 0.0000000 0.0000000   5.370638
#> 24 13.45884   5.081404 0.6931472 0.0000000 0.6931472   6.214608
#> 25 13.42985   5.159055 0.6931472 0.0000000 0.6931472   6.756932
#> 26 12.94801   4.330733 0.6931472 0.0000000 0.0000000   6.551080
#> 27 12.18075   3.871201 0.0000000      -Inf      -Inf   6.593045
#> 28 12.57764   4.189655 0.0000000      -Inf 0.0000000   6.613384
#> 29 12.51356   3.912023 0.0000000      -Inf 0.0000000   7.265430
#> 30 12.97154   4.110874 0.6931472      -Inf 0.0000000   5.135798
#> 31 13.70458   4.691348 1.0986123 0.0000000 0.0000000   6.272877
#> 32 13.01700   4.488636 0.6931472      -Inf 0.0000000   6.613384
#> 33 14.48334   5.973810 1.0986123 0.0000000 1.0986123   6.309918
#> 34 14.58098   6.359574 1.0986123 0.6931472 1.0986123   5.560682
#> 35 13.75364   5.204007 1.0986123 0.0000000 0.6931472   5.298317
#> 36 14.15198   5.568345 1.3862944 0.0000000 0.0000000   4.094345
#> 37 13.90169   5.323010 1.0986123      -Inf 1.0986123   6.142037
#> 38 14.05610   5.278115 1.0986123 1.0986123 0.6931472   6.413459
#> 39 14.84513   6.137727 1.0986123 1.0986123 1.0986123   6.380123
#> 40 14.40107   5.609472 1.0986123 1.0986123 1.3862944   4.941642
#> 41 14.15198   5.799093 1.3862944 0.6931472 0.6931472   6.484635
#> 42 14.91412   6.278521 1.3862944 1.0986123 1.3862944   6.056784
#> 43 13.99783   5.398163 1.0986123 1.0986123 0.6931472   6.408529
#> 44 13.59237   5.393628 1.0986123 0.0000000 0.0000000   6.907755
#> 45 13.76422   4.844187 0.6931472 0.0000000 0.0000000   4.094345
#> 46 14.53870   5.891644 1.0986123 1.0986123 1.3862944   5.736572
#> 47 14.09768   5.752573 1.0986123 1.0986123 1.0986123   6.396930
#> 48 13.65299   5.017280 1.0986123 0.0000000 0.6931472   6.492240
#> 49 14.31629   5.505332 1.0986123 1.0986123 1.0986123   5.726848
#> 50 13.38473   5.073422 1.0986123 0.0000000 0.0000000   4.787492
#> 
#> $sqrt
#>        valor area_total  quartos   suites garagens dist_b_mar
#> 1  1029.5630  18.708287 1.732051 1.000000 1.414214  26.832816
#> 2   714.1428  11.685889 1.732051 1.000000 1.000000  25.787594
#> 3   883.1761  12.836277 1.732051 1.000000 1.414214  20.371549
#> 4   741.6198  13.212873 1.732051 1.000000 1.000000  17.888544
#> 5   921.9544  11.090987 1.732051 1.000000 1.732051  29.916551
#> 6   547.7226   9.477869 1.414214 0.000000 1.000000  25.396850
#> 7   866.0254  13.190906 1.414214 1.000000 1.414214  29.325757
#> 8   806.2258  11.090537 1.732051 1.000000 1.000000  27.294688
#> 9   787.4008  11.000000 1.732051 1.000000 1.000000  27.294688
#> 10  860.2325  10.440307 1.732051 1.000000 1.000000  17.320508
#> 11  877.4964  13.038405 1.732051 1.000000 1.414214  24.289916
#> 12  824.6211  11.874342 1.732051 1.000000 1.000000  17.029386
#> 13  921.9544  13.190906 1.732051 1.000000 1.000000  21.563859
#> 14  648.0741  10.246951 1.732051 1.000000 0.000000   7.745967
#> 15  739.5945  11.313708 1.732051 1.000000 1.000000  27.294688
#> 16 1264.9111  12.767145 2.000000 1.414214 1.414214   9.486833
#> 17 1148.9125  15.165751 1.732051 1.000000 1.414214  14.662878
#> 18  784.2194  10.392305 1.732051 1.000000 1.000000  27.294688
#> 19  839.6428  13.190906 1.414214 1.000000 1.414214  30.000000
#> 20  646.5292   9.219544 1.000000 0.000000 1.000000  24.899799
#> 21  519.6152   8.426150 1.414214 0.000000 0.000000  37.148351
#> 22  646.5292  10.000000 1.000000 1.000000 1.000000  24.899799
#> 23  806.2258   9.486833 1.414214 1.000000 1.000000  14.662878
#> 24  836.6600  12.688578 1.414214 1.000000 1.414214  22.360680
#> 25  824.6211  13.190906 1.414214 1.000000 1.414214  29.325757
#> 26  648.0741   8.717798 1.414214 1.000000 1.000000  26.457513
#> 27  441.5880   6.928203 1.000000 0.000000 0.000000  27.018512
#> 28  538.5165   8.124038 1.000000 0.000000 1.000000  27.294688
#> 29  521.5362   7.071068 1.000000 0.000000 1.000000  37.815341
#> 30  655.7439   7.810250 1.414214 0.000000 1.000000  13.038405
#> 31  946.0444  10.440307 1.732051 1.000000 1.000000  23.021729
#> 32  670.8204   9.433981 1.414214 0.000000 1.000000  27.294688
#> 33 1396.4240  19.824228 1.732051 1.000000 1.732051  23.452079
#> 34 1466.2878  24.041631 1.732051 1.414214 1.732051  16.124515
#> 35  969.5360  13.490738 1.732051 1.000000 1.414214  14.142136
#> 36 1183.2160  16.186414 2.000000 1.000000 1.000000   7.745967
#> 37 1044.0307  14.317821 1.732051 0.000000 1.732051  21.563859
#> 38 1127.8298  14.000000 1.732051 1.732051 1.414214  24.698178
#> 39 1673.3201  21.517435 1.732051 1.732051 1.732051  24.289916
#> 40 1340.1492  16.522712 1.732051 1.732051 2.000000  11.832160
#> 41 1183.2160  18.165902 2.000000 1.414214 1.414214  25.592968
#> 42 1732.0508  23.086793 2.000000 1.732051 2.000000  20.663978
#> 43 1095.4451  14.866069 1.732051 1.732051 1.414214  24.637370
#> 44  894.4272  14.832397 1.732051 1.000000 1.000000  31.622777
#> 45  974.6794  11.269428 1.414214 1.000000 1.000000   7.745967
#> 46 1435.6183  19.026298 1.732051 1.732051 2.000000  17.606817
#> 47 1151.5207  17.748239 1.732051 1.732051 1.732051  24.494897
#> 48  921.9544  12.288206 1.732051 1.000000 1.414214  25.690465
#> 49 1284.5233  15.684387 1.732051 1.732051 1.732051  17.521415
#> 50  806.2258  12.638038 1.732051 1.000000 1.000000  10.954451
#> 

alltransf(centro_2015@data, 1:10, c("valor", "area_total"))

#> $identity
#>      valor area_total
#> 1  1060000     350.00
#> 2   510000     136.56
#> 3   780000     164.77
#> 4   550000     174.58
#> 5   850000     123.01
#> 6   300000      89.83
#> 7   750000     174.00
#> 8   650000     123.00
#> 9   620000     121.00
#> 10  740000     109.00
#> 
#> $rsqrt
#>           valor area_total
#> 1  0.0009712859 0.05345225
#> 2  0.0014002801 0.08557329
#> 3  0.0011322770 0.07790421
#> 4  0.0013483997 0.07568377
#> 5  0.0010846523 0.09016330
#> 6  0.0018257419 0.10550895
#> 7  0.0011547005 0.07580980
#> 8  0.0012403473 0.09016696
#> 9  0.0012700013 0.09090909
#> 10 0.0011624764 0.09578263
#> 
#> $log
#>       valor area_total
#> 1  13.87378   5.857933
#> 2  13.14217   4.916764
#> 3  13.56705   5.104551
#> 4  13.21767   5.162383
#> 5  13.65299   4.812266
#> 6  12.61154   4.497919
#> 7  13.52783   5.159055
#> 8  13.38473   4.812184
#> 9  13.33747   4.795791
#> 10 13.51441   4.691348
#> 
#> $sqrt
#>        valor area_total
#> 1  1029.5630  18.708287
#> 2   714.1428  11.685889
#> 3   883.1761  12.836277
#> 4   741.6198  13.212873
#> 5   921.9544  11.090987
#> 6   547.7226   9.477869
#> 7   866.0254  13.190906
#> 8   806.2258  11.090537
#> 9   787.4008  11.000000
#> 10  860.2325  10.440307
#> 

vars <- colnames(itacorubi_2015@data)
alltransf(itacorubi_2015@data, select = vars)

#> $identity
#>      valor area_total area_privativa quartos suites garagens
#> 1   577000        130          94.00       3      1        2
#> 2   480000        133          90.00       2      1        2
#> 3   650000        170         139.00       2      1        2
#> 4   450000        142          70.00       2      1        2
#> 5   440000        141          68.00       2      1        2
#> 6   815277        241         115.00       3      3        2
#> 7   436422        145          65.00       2      1        2
#> 8   881910        180         128.00       4      1        2
#> 9   498000        125          74.00       2      1        2
#> 10  430000         86          75.00       3      1        1
#> 11 1280000        301         168.00       4      1        3
#> 12  560000        205         146.00       3      2        1
#> 13  660000        145         103.00       3      1        2
#> 14  510600        156          65.00       2      1        2
#> 15  657000        129          94.00       3      1        2
#> 16  697000        176          88.00       3      1        2
#> 17  570000         93          70.00       2      1        2
#> 18  728030        219         105.00       3      1        2
#> 19  490000         86          74.00       2      1        1
#> 20  410000         84          74.00       3      1        2
#> 21  690000        286         194.00       3      1        3
#> 22  499500         86          72.00       3      1        2
#> 23  230000         53          38.00       1      1        1
#> 24  339200         70          49.00       1      1        1
#> 25 1490000        345         253.00       3      3        2
#> 26  490000        113          96.00       2      1        2
#> 27  780000        217          89.00       3      1        4
#> 28  470000        166         151.00       3      1        1
#> 29  440000        105          85.00       3      1        1
#> 30  390000        110          70.00       2      1        1
#> 31  365000         90          73.00       2      1        1
#> 32  595000        168         144.00       2      1        2
#> 33  635000        138         116.00       3      1        1
#> 34 1350000        225         180.00       4      2        3
#> 35  470000        187          77.00       2      2        2
#> 36  415000         90          69.00       2      1        2
#> 37  450000        100          80.00       2      1        2
#> 38  450000         95          80.00       2      1        1
#> 39  385000         78          78.00       2      1        1
#> 40  380000         75          75.00       2      1        1
#> 41  350000        102          69.00       2      1        1
#> 42  575000        118          93.00       3      1        2
#> 43  760000        143         120.00       3      1        2
#> 44 1050000        190         161.00       3      1        2
#> 45  259767         70          65.28       2      1        2
#> 46  900000        193         116.00       4      1        2
#> 47 1380000        301         165.00       4      2        3
#> 48  806388        207         122.00       4      1        2
#> 49 1487000        207         167.00       3      3        3
#> 50  864007        250         231.00       4      3        2
#> 51 1050000        180         166.00       3      3        2
#> 52  935422        246         122.00       4      2        3
#> 53  270000         64          50.00       1      0        1
#> 54  290000         67          51.00       1      0        1
#> 55  230000         52          43.00       1      0        1
#> 56  210000         71          45.00       1      0        1
#> 57  350000         60          48.00       1      1        1
#> 58  288000         80          56.00       2      0        1
#> 59  247000         65          53.00       2      0        1
#> 60  300000         78          66.00       2      0        1
#> 61  255000         65          52.00       2      0        1
#> 62  255000         66          58.00       3      0        1
#> 63  600000        168         144.00       2      1        2
#> 
#> $rsqrt
#>           valor area_total area_privativa   quartos    suites  garagens
#> 1  0.0013164734 0.08770580     0.10314212 0.5773503 1.0000000 0.7071068
#> 2  0.0014433757 0.08671100     0.10540926 0.7071068 1.0000000 0.7071068
#> 3  0.0012403473 0.07669650     0.08481889 0.7071068 1.0000000 0.7071068
#> 4  0.0014907120 0.08391814     0.11952286 0.7071068 1.0000000 0.7071068
#> 5  0.0015075567 0.08421519     0.12126781 0.7071068 1.0000000 0.7071068
#> 6  0.0011075094 0.06441566     0.09325048 0.5773503 0.5773503 0.7071068
#> 7  0.0015137240 0.08304548     0.12403473 0.7071068 1.0000000 0.7071068
#> 8  0.0010648486 0.07453560     0.08838835 0.5000000 1.0000000 0.7071068
#> 9  0.0014170505 0.08944272     0.11624764 0.7071068 1.0000000 0.7071068
#> 10 0.0015249857 0.10783277     0.11547005 0.5773503 1.0000000 1.0000000
#> 11 0.0008838835 0.05763904     0.07715167 0.5000000 1.0000000 0.5773503
#> 12 0.0013363062 0.06984303     0.08276059 0.5773503 0.7071068 1.0000000
#> 13 0.0012309149 0.08304548     0.09853293 0.5773503 1.0000000 0.7071068
#> 14 0.0013994571 0.08006408     0.12403473 0.7071068 1.0000000 0.7071068
#> 15 0.0012337220 0.08804509     0.10314212 0.5773503 1.0000000 0.7071068
#> 16 0.0011977981 0.07537784     0.10660036 0.5773503 1.0000000 0.7071068
#> 17 0.0013245324 0.10369517     0.11952286 0.7071068 1.0000000 0.7071068
#> 18 0.0011719939 0.06757374     0.09759001 0.5773503 1.0000000 0.7071068
#> 19 0.0014285714 0.10783277     0.11624764 0.7071068 1.0000000 1.0000000
#> 20 0.0015617376 0.10910895     0.11624764 0.5773503 1.0000000 0.7071068
#> 21 0.0012038585 0.05913124     0.07179582 0.5773503 1.0000000 0.5773503
#> 22 0.0014149212 0.10783277     0.11785113 0.5773503 1.0000000 0.7071068
#> 23 0.0020851441 0.13736056     0.16222142 1.0000000 1.0000000 1.0000000
#> 24 0.0017170070 0.11952286     0.14285714 1.0000000 1.0000000 1.0000000
#> 25 0.0008192319 0.05383819     0.06286946 0.5773503 0.5773503 0.7071068
#> 26 0.0014285714 0.09407209     0.10206207 0.7071068 1.0000000 0.7071068
#> 27 0.0011322770 0.06788442     0.10599979 0.5773503 1.0000000 0.5000000
#> 28 0.0014586499 0.07761505     0.08137885 0.5773503 1.0000000 1.0000000
#> 29 0.0015075567 0.09759001     0.10846523 0.5773503 1.0000000 1.0000000
#> 30 0.0016012815 0.09534626     0.11952286 0.7071068 1.0000000 1.0000000
#> 31 0.0016552118 0.10540926     0.11704115 0.7071068 1.0000000 1.0000000
#> 32 0.0012964074 0.07715167     0.08333333 0.7071068 1.0000000 0.7071068
#> 33 0.0012549116 0.08512565     0.09284767 0.5773503 1.0000000 1.0000000
#> 34 0.0008606630 0.06666667     0.07453560 0.5000000 0.7071068 0.5773503
#> 35 0.0014586499 0.07312724     0.11396058 0.7071068 0.7071068 0.7071068
#> 36 0.0015523011 0.10540926     0.12038585 0.7071068 1.0000000 0.7071068
#> 37 0.0014907120 0.10000000     0.11180340 0.7071068 1.0000000 0.7071068
#> 38 0.0014907120 0.10259784     0.11180340 0.7071068 1.0000000 1.0000000
#> 39 0.0016116459 0.11322770     0.11322770 0.7071068 1.0000000 1.0000000
#> 40 0.0016222142 0.11547005     0.11547005 0.7071068 1.0000000 1.0000000
#> 41 0.0016903085 0.09901475     0.12038585 0.7071068 1.0000000 1.0000000
#> 42 0.0013187609 0.09205746     0.10369517 0.5773503 1.0000000 0.7071068
#> 43 0.0011470787 0.08362420     0.09128709 0.5773503 1.0000000 0.7071068
#> 44 0.0009759001 0.07254763     0.07881104 0.5773503 1.0000000 0.7071068
#> 45 0.0019620407 0.11952286     0.12376844 0.7071068 1.0000000 0.7071068
#> 46 0.0010540926 0.07198158     0.09284767 0.5000000 1.0000000 0.7071068
#> 47 0.0008512565 0.05763904     0.07784989 0.5000000 0.7071068 0.5773503
#> 48 0.0011135968 0.06950480     0.09053575 0.5000000 1.0000000 0.7071068
#> 49 0.0008200579 0.06950480     0.07738232 0.5773503 0.5773503 0.5773503
#> 50 0.0010758243 0.06324555     0.06579517 0.5000000 0.5773503 0.7071068
#> 51 0.0009759001 0.07453560     0.07761505 0.5773503 0.5773503 0.7071068
#> 52 0.0010339421 0.06375767     0.09053575 0.5000000 0.7071068 0.5773503
#> 53 0.0019245009 0.12500000     0.14142136 1.0000000       Inf 1.0000000
#> 54 0.0018569534 0.12216944     0.14002801 1.0000000       Inf 1.0000000
#> 55 0.0020851441 0.13867505     0.15249857 1.0000000       Inf 1.0000000
#> 56 0.0021821789 0.11867817     0.14907120 1.0000000       Inf 1.0000000
#> 57 0.0016903085 0.12909944     0.14433757 1.0000000 1.0000000 1.0000000
#> 58 0.0018633900 0.11180340     0.13363062 0.7071068       Inf 1.0000000
#> 59 0.0020121091 0.12403473     0.13736056 0.7071068       Inf 1.0000000
#> 60 0.0018257419 0.11322770     0.12309149 0.7071068       Inf 1.0000000
#> 61 0.0019802951 0.12403473     0.13867505 0.7071068       Inf 1.0000000
#> 62 0.0019802951 0.12309149     0.13130643 0.5773503       Inf 1.0000000
#> 63 0.0012909944 0.07715167     0.08333333 0.7071068 1.0000000 0.7071068
#> 
#> $log
#>       valor area_total area_privativa   quartos    suites  garagens
#> 1  13.26560   4.867534       4.543295 1.0986123 0.0000000 0.6931472
#> 2  13.08154   4.890349       4.499810 0.6931472 0.0000000 0.6931472
#> 3  13.38473   5.135798       4.934474 0.6931472 0.0000000 0.6931472
#> 4  13.01700   4.955827       4.248495 0.6931472 0.0000000 0.6931472
#> 5  12.99453   4.948760       4.219508 0.6931472 0.0000000 0.6931472
#> 6  13.61128   5.484797       4.744932 1.0986123 1.0986123 0.6931472
#> 7  12.98636   4.976734       4.174387 0.6931472 0.0000000 0.6931472
#> 8  13.68985   5.192957       4.852030 1.3862944 0.0000000 0.6931472
#> 9  13.11836   4.828314       4.304065 0.6931472 0.0000000 0.6931472
#> 10 12.97154   4.454347       4.317488 1.0986123 0.0000000 0.0000000
#> 11 14.06237   5.707110       5.123964 1.3862944 0.0000000 1.0986123
#> 12 13.23569   5.323010       4.983607 1.0986123 0.6931472 0.0000000
#> 13 13.40000   4.976734       4.634729 1.0986123 0.0000000 0.6931472
#> 14 13.14334   5.049856       4.174387 0.6931472 0.0000000 0.6931472
#> 15 13.39544   4.859812       4.543295 1.0986123 0.0000000 0.6931472
#> 16 13.45454   5.170484       4.477337 1.0986123 0.0000000 0.6931472
#> 17 13.25339   4.532599       4.248495 0.6931472 0.0000000 0.6931472
#> 18 13.49810   5.389072       4.653960 1.0986123 0.0000000 0.6931472
#> 19 13.10216   4.454347       4.304065 0.6931472 0.0000000 0.0000000
#> 20 12.92391   4.430817       4.304065 1.0986123 0.0000000 0.6931472
#> 21 13.44445   5.655992       5.267858 1.0986123 0.0000000 1.0986123
#> 22 13.12136   4.454347       4.276666 1.0986123 0.0000000 0.6931472
#> 23 12.34583   3.970292       3.637586 0.0000000 0.0000000 0.0000000
#> 24 12.73435   4.248495       3.891820 0.0000000 0.0000000 0.0000000
#> 25 14.21429   5.843544       5.533389 1.0986123 1.0986123 0.6931472
#> 26 13.10216   4.727388       4.564348 0.6931472 0.0000000 0.6931472
#> 27 13.56705   5.379897       4.488636 1.0986123 0.0000000 1.3862944
#> 28 13.06049   5.111988       5.017280 1.0986123 0.0000000 0.0000000
#> 29 12.99453   4.653960       4.442651 1.0986123 0.0000000 0.0000000
#> 30 12.87390   4.700480       4.248495 0.6931472 0.0000000 0.0000000
#> 31 12.80765   4.499810       4.290459 0.6931472 0.0000000 0.0000000
#> 32 13.29632   5.123964       4.969813 0.6931472 0.0000000 0.6931472
#> 33 13.36138   4.927254       4.753590 1.0986123 0.0000000 0.0000000
#> 34 14.11562   5.416100       5.192957 1.3862944 0.6931472 1.0986123
#> 35 13.06049   5.231109       4.343805 0.6931472 0.6931472 0.6931472
#> 36 12.93603   4.499810       4.234107 0.6931472 0.0000000 0.6931472
#> 37 13.01700   4.605170       4.382027 0.6931472 0.0000000 0.6931472
#> 38 13.01700   4.553877       4.382027 0.6931472 0.0000000 0.0000000
#> 39 12.86100   4.356709       4.356709 0.6931472 0.0000000 0.0000000
#> 40 12.84793   4.317488       4.317488 0.6931472 0.0000000 0.0000000
#> 41 12.76569   4.624973       4.234107 0.6931472 0.0000000 0.0000000
#> 42 13.26213   4.770685       4.532599 1.0986123 0.0000000 0.6931472
#> 43 13.54107   4.962845       4.787492 1.0986123 0.0000000 0.6931472
#> 44 13.86430   5.247024       5.081404 1.0986123 0.0000000 0.6931472
#> 45 12.46754   4.248495       4.178686 0.6931472 0.0000000 0.6931472
#> 46 13.71015   5.262690       4.753590 1.3862944 0.0000000 0.6931472
#> 47 14.13759   5.707110       5.105945 1.3862944 0.6931472 1.0986123
#> 48 13.60032   5.332719       4.804021 1.3862944 0.0000000 0.6931472
#> 49 14.21227   5.332719       5.117994 1.0986123 1.0986123 1.0986123
#> 50 13.66934   5.521461       5.442418 1.3862944 1.0986123 0.6931472
#> 51 13.86430   5.192957       5.111988 1.0986123 1.0986123 0.6931472
#> 52 13.74875   5.505332       4.804021 1.3862944 0.6931472 1.0986123
#> 53 12.50618   4.158883       3.912023 0.0000000      -Inf 0.0000000
#> 54 12.57764   4.204693       3.931826 0.0000000      -Inf 0.0000000
#> 55 12.34583   3.951244       3.761200 0.0000000      -Inf 0.0000000
#> 56 12.25486   4.262680       3.806662 0.0000000      -Inf 0.0000000
#> 57 12.76569   4.094345       3.871201 0.0000000 0.0000000 0.0000000
#> 58 12.57072   4.382027       4.025352 0.6931472      -Inf 0.0000000
#> 59 12.41714   4.174387       3.970292 0.6931472      -Inf 0.0000000
#> 60 12.61154   4.356709       4.189655 0.6931472      -Inf 0.0000000
#> 61 12.44902   4.174387       3.951244 0.6931472      -Inf 0.0000000
#> 62 12.44902   4.189655       4.060443 1.0986123      -Inf 0.0000000
#> 63 13.30468   5.123964       4.969813 0.6931472 0.0000000 0.6931472
#> 
#> $sqrt
#>        valor area_total area_privativa  quartos   suites garagens
#> 1   759.6052  11.401754       9.695360 1.732051 1.000000 1.414214
#> 2   692.8203  11.532563       9.486833 1.414214 1.000000 1.414214
#> 3   806.2258  13.038405      11.789826 1.414214 1.000000 1.414214
#> 4   670.8204  11.916375       8.366600 1.414214 1.000000 1.414214
#> 5   663.3250  11.874342       8.246211 1.414214 1.000000 1.414214
#> 6   902.9269  15.524175      10.723805 1.732051 1.732051 1.414214
#> 7   660.6224  12.041595       8.062258 1.414214 1.000000 1.414214
#> 8   939.1006  13.416408      11.313708 2.000000 1.000000 1.414214
#> 9   705.6912  11.180340       8.602325 1.414214 1.000000 1.414214
#> 10  655.7439   9.273618       8.660254 1.732051 1.000000 1.000000
#> 11 1131.3708  17.349352      12.961481 2.000000 1.000000 1.732051
#> 12  748.3315  14.317821      12.083046 1.732051 1.414214 1.000000
#> 13  812.4038  12.041595      10.148892 1.732051 1.000000 1.414214
#> 14  714.5628  12.489996       8.062258 1.414214 1.000000 1.414214
#> 15  810.5554  11.357817       9.695360 1.732051 1.000000 1.414214
#> 16  834.8653  13.266499       9.380832 1.732051 1.000000 1.414214
#> 17  754.9834   9.643651       8.366600 1.414214 1.000000 1.414214
#> 18  853.2467  14.798649      10.246951 1.732051 1.000000 1.414214
#> 19  700.0000   9.273618       8.602325 1.414214 1.000000 1.000000
#> 20  640.3124   9.165151       8.602325 1.732051 1.000000 1.414214
#> 21  830.6624  16.911535      13.928388 1.732051 1.000000 1.732051
#> 22  706.7531   9.273618       8.485281 1.732051 1.000000 1.414214
#> 23  479.5832   7.280110       6.164414 1.000000 1.000000 1.000000
#> 24  582.4088   8.366600       7.000000 1.000000 1.000000 1.000000
#> 25 1220.6556  18.574176      15.905974 1.732051 1.732051 1.414214
#> 26  700.0000  10.630146       9.797959 1.414214 1.000000 1.414214
#> 27  883.1761  14.730920       9.433981 1.732051 1.000000 2.000000
#> 28  685.5655  12.884099      12.288206 1.732051 1.000000 1.000000
#> 29  663.3250  10.246951       9.219544 1.732051 1.000000 1.000000
#> 30  624.4998  10.488088       8.366600 1.414214 1.000000 1.000000
#> 31  604.1523   9.486833       8.544004 1.414214 1.000000 1.000000
#> 32  771.3624  12.961481      12.000000 1.414214 1.000000 1.414214
#> 33  796.8689  11.747340      10.770330 1.732051 1.000000 1.000000
#> 34 1161.8950  15.000000      13.416408 2.000000 1.414214 1.732051
#> 35  685.5655  13.674794       8.774964 1.414214 1.414214 1.414214
#> 36  644.2049   9.486833       8.306624 1.414214 1.000000 1.414214
#> 37  670.8204  10.000000       8.944272 1.414214 1.000000 1.414214
#> 38  670.8204   9.746794       8.944272 1.414214 1.000000 1.000000
#> 39  620.4837   8.831761       8.831761 1.414214 1.000000 1.000000
#> 40  616.4414   8.660254       8.660254 1.414214 1.000000 1.000000
#> 41  591.6080  10.099505       8.306624 1.414214 1.000000 1.000000
#> 42  758.2875  10.862780       9.643651 1.732051 1.000000 1.414214
#> 43  871.7798  11.958261      10.954451 1.732051 1.000000 1.414214
#> 44 1024.6951  13.784049      12.688578 1.732051 1.000000 1.414214
#> 45  509.6734   8.366600       8.079604 1.414214 1.000000 1.414214
#> 46  948.6833  13.892444      10.770330 2.000000 1.000000 1.414214
#> 47 1174.7340  17.349352      12.845233 2.000000 1.414214 1.732051
#> 48  897.9911  14.387495      11.045361 2.000000 1.000000 1.414214
#> 49 1219.4261  14.387495      12.922848 1.732051 1.732051 1.732051
#> 50  929.5198  15.811388      15.198684 2.000000 1.732051 1.414214
#> 51 1024.6951  13.416408      12.884099 1.732051 1.732051 1.414214
#> 52  967.1722  15.684387      11.045361 2.000000 1.414214 1.732051
#> 53  519.6152   8.000000       7.071068 1.000000 0.000000 1.000000
#> 54  538.5165   8.185353       7.141428 1.000000 0.000000 1.000000
#> 55  479.5832   7.211103       6.557439 1.000000 0.000000 1.000000
#> 56  458.2576   8.426150       6.708204 1.000000 0.000000 1.000000
#> 57  591.6080   7.745967       6.928203 1.000000 1.000000 1.000000
#> 58  536.6563   8.944272       7.483315 1.414214 0.000000 1.000000
#> 59  496.9909   8.062258       7.280110 1.414214 0.000000 1.000000
#> 60  547.7226   8.831761       8.124038 1.414214 0.000000 1.000000
#> 61  504.9752   8.062258       7.211103 1.414214 0.000000 1.000000
#> 62  504.9752   8.124038       7.615773 1.732051 0.000000 1.000000
#> 63  774.5967  12.961481      12.000000 1.414214 1.000000 1.414214
#>