Summary method for class bestfit

summary method for class bestfit produce result summary containing the formula for the best (chosen) fit and the summary.lm for that fit.

# S3 method for bestfit
summary(object, fit = 1, subset, ...)

Arguments

object
an object of class bestfit.

fit
the number of the chosen fit from the combination matrix (defaults for the best fit found with bestfit).

subset
a specification of the rows to be used: defaults to all rows. This can be any valid indexing vector (see [.data.frame) for the rows of data or if that is not supplied, a data frame made up of the variables used in formula.

…
further arguments passed to lm.

Value

Returns the call for the bestfit function, the best (chosen) fit number, the lm formula and the lm fit summary for the best (chosen) fit transformations found by bestfit.

Examples

best_fit <- bestfit(valor ~ ., data = centro_2015@data)
summary(best_fit)

#> Call:
#> bestfit.formula(formula = valor ~ ., data = centro_2015@data)
#> 
#> Best (Chosen) Transformations:
#>     id valor area_total quartos   suites garagens dist_b_mar    adj_R2
#> 443  1 rsqrt       sqrt   rsqrt identity     sqrt      rsqrt 0.9480455
#> 
#> Best (Chosen) fit LM summary:
#> 
#> Call:
#> lm(formula = "rsqrt(valor) ~ sqrt(area_total) + rsqrt(quartos) + identity(suites) + sqrt(garagens) + rsqrt(dist_b_mar) + (padrao)", 
#>     data = centro_2015@data, subset = NULL)
#> 
#> Residuals:
#>        Min         1Q     Median         3Q        Max 
#> -2.144e-04 -5.344e-05  8.870e-07  4.272e-05  1.729e-04 
#> 
#> Coefficients:
#>                     Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)        1.780e-03  1.301e-04  13.681  < 2e-16 ***
#> sqrt(area_total)  -2.279e-05  5.295e-06  -4.304 9.81e-05 ***
#> rsqrt(quartos)     6.561e-04  1.269e-04   5.169 6.14e-06 ***
#> identity(suites)  -4.240e-05  2.060e-05  -2.058   0.0459 *  
#> sqrt(garagens)    -2.711e-04  4.426e-05  -6.125 2.62e-07 ***
#> rsqrt(dist_b_mar) -2.628e-03  5.099e-04  -5.154 6.45e-06 ***
#> padraomedio       -2.214e-04  4.586e-05  -4.828 1.85e-05 ***
#> padraoalto        -2.576e-04  4.605e-05  -5.595 1.52e-06 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> Residual standard error: 8.428e-05 on 42 degrees of freedom
#>   (3 observations deleted due to missingness)
#> Multiple R-squared:  0.9555,	Adjusted R-squared:  0.948 
#> F-statistic: 128.7 on 7 and 42 DF,  p-value: < 2.2e-16
#> 
#> NBR-14.653-2 check:
#> Minimum number of market data:
#> [1] "n = 53 >= 42 --> Grau III"
#> Max significance level allowed for each predictor:
#> [1] "t máximo = 4.59 %  < 10% --> Grau III"
#> Max significance level allowed for F-test:
#> [1] "p-valor F = 2.79e-24 % < 1% --> Grau III"

summary(best_fit, fit = 2)

#> Call:
#> bestfit.formula(formula = valor ~ ., data = centro_2015@data)
#> 
#> Best (Chosen) Transformations:
#>     id valor area_total quartos   suites garagens dist_b_mar    adj_R2
#> 395  2 rsqrt   identity   rsqrt identity     sqrt      rsqrt 0.9477222
#> 
#> Best (Chosen) fit LM summary:
#> 
#> Call:
#> lm(formula = "rsqrt(valor) ~ identity(area_total) + rsqrt(quartos) + identity(suites) + sqrt(garagens) + rsqrt(dist_b_mar) + (padrao)", 
#>     data = centro_2015@data, subset = NULL)
#> 
#> Residuals:
#>        Min         1Q     Median         3Q        Max 
#> -2.012e-04 -5.206e-05 -1.442e-06  4.426e-05  1.674e-04 
#> 
#> Coefficients:
#>                        Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)           1.608e-03  1.176e-04  13.672  < 2e-16 ***
#> identity(area_total) -6.702e-07  1.573e-07  -4.261 0.000113 ***
#> rsqrt(quartos)        7.006e-04  1.246e-04   5.623 1.38e-06 ***
#> identity(suites)     -4.472e-05  2.052e-05  -2.180 0.034933 *  
#> sqrt(garagens)       -2.779e-04  4.377e-05  -6.350 1.25e-07 ***
#> rsqrt(dist_b_mar)    -2.668e-03  5.112e-04  -5.220 5.19e-06 ***
#> padraomedio          -2.428e-04  4.633e-05  -5.241 4.85e-06 ***
#> padraoalto           -2.770e-04  4.617e-05  -6.000 3.97e-07 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> Residual standard error: 8.454e-05 on 42 degrees of freedom
#>   (3 observations deleted due to missingness)
#> Multiple R-squared:  0.9552,	Adjusted R-squared:  0.9477 
#> F-statistic: 127.9 on 7 and 42 DF,  p-value: < 2.2e-16
#> 
#> NBR-14.653-2 check:
#> Minimum number of market data:
#> [1] "n = 53 >= 42 --> Grau III"
#> Max significance level allowed for each predictor:
#> [1] "t máximo = 3.49 %  < 10% --> Grau III"
#> Max significance level allowed for F-test:
#> [1] "p-valor F = 3.17e-24 % < 1% --> Grau III"